Implemented global search for games under globalGames.controllers.js (Haven't tested it) And trying to make sure it is not case sensitive.
Also, started working on the implementation of profile update. So far only copied the
function above it to fetch the data. The data shall be put in boxes spawned by front end
and accept whatever new chanegs are done. Not everything can be updated, like userID for
example. Also, consider having a cancel button to not do any changes.